[CORE] Backport crypto lib

+++ src/crypto/sha1.h
@@ -1,4 +1,4 @@
-// Copyright (c) 2014 The Bitcoin developers
+// Copyright (c) 2014-2016 The Bitcoin Core developers
 // Distributed under the MIT software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
@@ -12,17 +12,17 @@
 class CSHA1
 {
 private:
-    uint32_t s[5];
-    unsigned char buf[64];
-    size_t bytes;
+  uint32_t s[5];
+  unsigned char buf[64];
+  uint64_t bytes;
 
 public:
-    static const size_t OUTPUT_SIZE = 20;
+  static const size_t OUTPUT_SIZE = 20;
 
-    CSHA1();
-    CSHA1& Write(const unsigned char* data, size_t len);
-    void Finalize(unsigned char hash[OUTPUT_SIZE]);
-    CSHA1& Reset();
+  CSHA1();
+  CSHA1& Write(const unsigned char* data, size_t len);
+  void Finalize(unsigned char hash[OUTPUT_SIZE]);
+  CSHA1& Reset();
 };
 
 #endif // BITCOIN_CRYPTO_SHA1_H
